Network aggregations are working
