+++ data/.gitignore
@@ -1,2 +1,3 @@
 *.zip
 *.h5
+*.csv
